requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.31.1

# Bleeding edge Django
django>=2.0.5, <2.2
django-crispy-forms==1.7.2
# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Forms

django-crispy-forms==1.7.2

# Models
django-model-utils==3.1.2

# Images
Pillow==5.1.0

# Password storage
argon2-cffi==18.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.36.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.4

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2018.4

# Redis support
redis>=2.10.5
django-redis==4.9.0


celery==4.1.1

# Your custom requirements go here

# mysqlclient==1.3.12

django-rest-auth==0.9.3 # intergate with rest framework and allauth package
drf-writable-nested==0.4.2

# django-import-export==1.0.0
pandas==0.23.0

# djangorestframework
djangorestframework==3.8.2
django-cors-headers==2.2.0
djangorestframework-jwt==1.11.0
django-filter==1.1.0

# for celery
flower==0.9.2

requests==2.18.4

channels==2.1.1
channels_redis==2.2.1
